improve create page loading experience (closes #146)
