+++ pages/create/index.tsx
@@ -1,6 +1,7 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { AppContext } from '../../contexts/appContext';
 import Dimensions from '../../constants/dimensions';
+import Level from '../../models/db/level';
 import LevelTable from '../../components/levelTable';
 import Page from '../../components/page';
 import World from '../../models/db/world';
@@ -10,6 +11,7 @@ import useUser from '../../hooks/useUser';
 
 export default function Create() {
   const { isLoading, user } = useUser();
+  const [levels, setLevels] = useState<Level[]>();
   const router = useRouter();
   const { setIsLoading } = useContext(AppContext);
   const [worlds, setWorlds] = useState<World[]>();
@@ -19,6 +21,21 @@ export default function Create() {
       router.replace('/');
     }
   }, [isLoading, router, user]);
+
+  const getLevels = useCallback(() => {
+    fetch('/api/levels', {
+      method: 'GET',
+    }).then(async res => {
+      if (res.status === 200) {
+        setLevels(await res.json());
+      } else {
+        throw res.text();
+      }
+    }).catch(err => {
+      console.error(err);
+      alert('Error fetching levels');
+    });
+  }, []);
 
   const getWorlds = useCallback(() => {
     fetch('/api/worlds', {
@@ -36,16 +53,16 @@ export default function Create() {
   }, []);
 
   useEffect(() => {
+    getLevels();
+  }, [getLevels]);
+
+  useEffect(() => {
     getWorlds();
   }, [getWorlds]);
 
   useEffect(() => {
-    setIsLoading(!worlds);
-  }, [setIsLoading, worlds]);
-
-  if (!worlds) {
-    return null;
-  }
+    setIsLoading(!levels || !worlds);
+  }, [levels, setIsLoading, worlds]);
 
   return (
     <Page title={'Create'}>
@@ -61,7 +78,7 @@ export default function Create() {
           Welcome to the Create page! Here you can create worlds and levels. After creating a level, click on its name to start editing. Once you have finished desgining your level, click the &apos;Test&apos; button to set the level&apos;s least moves, then click publish to make your level available for everyone to play. When publishing a level you can decide if you want it to exist in any of your worlds. Note that a world will not appear in the catalog until it has at least one published level. You can unpublish or delete a level at any time.
         </div>
         <WorldTable getWorlds={getWorlds} worlds={worlds} />
-        <LevelTable worlds={worlds} />
+        <LevelTable getLevels={getLevels} levels={levels} worlds={worlds} />
       </>
     </Page>
   );
